feat: 인증 인가 기능 구현

+++ src/main/java/woowacourse/shoppingcart/controller/CustomerController.java
@@ -4,16 +4,18 @@ import java.net.URI;
 import javax.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import woowacourse.auth.support.AuthenticationPrincipal;
+import woowacourse.auth.support.LoginCustomer;
 import woowacourse.shoppingcart.dto.CustomerDto;
 import woowacourse.shoppingcart.dto.SignUpDto;
 import woowacourse.shoppingcart.dto.UpdateCustomerDto;
+import woowacourse.shoppingcart.exception.ForbiddenException;
 import woowacourse.shoppingcart.service.CustomerService;
 
 @RestController
@@ -34,14 +36,25 @@ public class CustomerController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<CustomerDto> updateCustomer(@PathVariable final Long id, @RequestBody UpdateCustomerDto updateCustomerDto) {
+    public ResponseEntity<CustomerDto> updateCustomer(@PathVariable final Long id,
+                                                      @RequestBody UpdateCustomerDto updateCustomerDto,
+                                                      @AuthenticationPrincipal LoginCustomer loginCustomer) {
+        checkAuthorization(id, loginCustomer);
         final CustomerDto customerDto = customerService.updateCustomer(id, updateCustomerDto);
 
         return ResponseEntity.ok(customerDto);
     }
 
+    private void checkAuthorization(final Long id, final LoginCustomer loginCustomer) {
+        if (!loginCustomer.getId().equals(id)) {
+            throw new ForbiddenException();
+        }
+    }
+
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteCustomer(@PathVariable final Long id) {
+    public ResponseEntity<Void> deleteCustomer(@PathVariable final Long id,
+                                               @AuthenticationPrincipal LoginCustomer loginCustomer) {
+        checkAuthorization(id, loginCustomer);
         customerService.deleteCustomer(id);
 
         return ResponseEntity.noContent().build();
